Update
-Changed some class and id names in the html file.

-Gathering all the fields in javascript with queryselector.

- Created checkEmptyField function that checks first if theres a error that already exist and searches the index of the error array. If not found it's set to -1.
Then it checks the veld type. if checkbox it checks if its checked or not and pushes the error into the array. if checked to removed it out of the array.
If veld type is a form where the value is empty then we do the same as above.

- Created Validate function that checks what attribute the form has and then we use the checkEmptyField.
Then we check if the array is empty or not . If empty then the alert will be succes and appears green. If not then we show all the errors in the array and let it appear red.